requirements.txt
customtkinter>=5.2.0
pywebview>=4.4.0
